Add media file size limits to upload per instance config

+++ husky/app/src/main/java/com/keylesspalace/tusky/components/instance/data/models/data/MstdnConfiguration.kt
@@ -18,7 +18,7 @@ data class MstdnConfigStatuses(
 
 data class MstdnStatusesMediaAttachment(
     @SerializedName("image_size_limit")
-    val imageSizeLimit: Int?,
+    val imageSizeLimit: Long?,
     @SerializedName("video_size_limit")
-    val videoSizeLimit: Int?
+    val videoSizeLimit: Long?
 )
